Abstracts ImperativeDialog
